Potential fix for code scanning alert no. 9: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to prevent overflow. We can achieve this by casting one of the operands to size_t before performing the multiplication. This will ensure that the multiplication is done using the larger type, thus avoiding overflow.

Specifically, we will cast newMaxRows to size_t before performing the multiplication on line 218. This change will ensure that the multiplication is done using size_t, which is typically larger than int and can hold larger values.

+++ libicu4c-64-2/common/propsvec.cpp
@@ -215,7 +215,7 @@ upvec_setValue(UPropsVectors *pv,
                 *pErrorCode=U_INTERNAL_PROGRAM_ERROR;
                 return;
             }
-            newVectors=(uint32_t *)uprv_malloc(newMaxRows*columns*4);
+            newVectors=(uint32_t *)uprv_malloc((size_t)newMaxRows*columns*4);
             if(newVectors==NULL) {
                 *pErrorCode=U_MEMORY_ALLOCATION_ERROR;
                 return;
